+++ frontend/src/hooks/TreeMap/useTreeMap.ts
@@ -1,14 +1,9 @@
 import { useEffect, useRef, useState } from 'react';
 import { DEFAULT_LATITUDE, DEFAULT_LONGITUDE } from '@/constants/map';
-import treeImage from '@/assets/TREE_01.png';
 
 const { kakao } = window;
 const DEFAULT_ZOOM_LEVEL = 3;
 const USER_LOCATION_STORAGE_KEY = 'userLocation';
-
-const MARKER_IMAGE: Record<string, string> = {
-  TREE_01: treeImage,
-};
 
 interface Coordinates {
   latitude: number;
@@ -54,76 +49,42 @@ const calculateMapBounds = (map: typeof kakao.maps.Map | null): Bounds => {
   };
 };
 
-const createMarkerImage = (imageCode: string) => {
-  return new kakao.maps.MarkerImage(MARKER_IMAGE[imageCode], new kakao.maps.Size(50, 55), {
-    offset: new kakao.maps.Point(25, 55),
-  });
-};
-
 const useTreeMap = () => {
   const mapRef = useRef<HTMLDivElement | null>(null);
   const [map, setMap] = useState<typeof kakao.maps.Map | null>(null);
   const [zoom, setZoom] = useState<number>(DEFAULT_ZOOM_LEVEL);
   const [centerPosition, setCenterPosition] = useState<Coordinates>(locationStorage.get);
-  const [bounds, setBounds] = useState<Bounds>(() => calculateMapBounds(map));
-  const currentMarkers = useRef<(typeof kakao.maps.Marker | typeof kakao.maps.CustomOverlay)[]>([]);
+  const [bounds, setBounds] = useState<Bounds>(() => calculateMapBounds(null));
 
-  const addMarker = (
-    targetMap: typeof kakao.maps.Map,
-    latitude: number,
-    longitude: number,
-    imageCode: string,
-    onClick?: () => void,
-  ) => {
-    const position = new kakao.maps.LatLng(latitude, longitude);
-    const image = createMarkerImage(imageCode);
+  useEffect(() => {
+    if (!mapRef.current || !kakao?.maps) return;
 
-    const marker = new kakao.maps.Marker({
-      position,
-      image,
-      clickable: true,
+    const initialCoordinates = locationStorage.get();
+    const mapInstance = new kakao.maps.Map(mapRef.current, {
+      center: new kakao.maps.LatLng(initialCoordinates.latitude, initialCoordinates.longitude),
+      level: DEFAULT_ZOOM_LEVEL,
     });
 
-    if (onClick) {
-      kakao.maps.event.addListener(marker, 'click', onClick);
-    }
+    setMap(mapInstance);
+    setZoom(mapInstance.getLevel());
+    setBounds(calculateMapBounds(mapInstance));
 
-    marker.setMap(targetMap);
-    currentMarkers.current.push(marker);
-  };
+    const handleZoomChange = () => {
+      setZoom(mapInstance.getLevel());
+      setBounds(calculateMapBounds(mapInstance));
+    };
 
-  const addCustomOverlay = (
-    targetMap: typeof kakao.maps.Map,
-    latitude: number,
-    longitude: number,
-    content: string,
-    onClick?: () => void,
-  ) => {
-    const position = new kakao.maps.LatLng(latitude, longitude);
+    kakao.maps.event.addListener(mapInstance, 'zoom_changed', handleZoomChange);
 
-    const overlay = new kakao.maps.CustomOverlay({
-      position,
-      content,
-      clickable: true,
-    });
-
-    if (onClick) {
-      kakao.maps.event.addListener(overlay, 'click', onClick);
-    }
-
-    overlay.setMap(targetMap);
-    currentMarkers.current.push(overlay);
-  };
-
-  const clearMarkers = () => {
-    currentMarkers.current.forEach((marker) => marker.setMap(null));
-    currentMarkers.current = [];
-  };
+    return () => {
+      kakao.maps.event.removeListener(mapInstance, 'zoom_changed', handleZoomChange);
+    };
+  }, []);
 
   const updateCenterPosition = () => {
-    const center = map?.getCenter();
-    if (!center) return;
+    if (!map) return;
 
+    const center = map.getCenter();
     const coordinates: Coordinates = {
       latitude: center.getLat(),
       longitude: center.getLng(),
@@ -135,8 +96,7 @@ const useTreeMap = () => {
 
   const updateBounds = () => {
     if (!map) return;
-    const newBounds = calculateMapBounds(map);
-    setBounds(newBounds);
+    setBounds(calculateMapBounds(map));
   };
 
   const updatePosition = () => {
@@ -144,27 +104,19 @@ const useTreeMap = () => {
     updateCenterPosition();
   };
 
-  useEffect(() => {
-    const initialCoordinates = locationStorage.get();
+  const zoomIn = () => {
+    if (!map) return;
 
-    const initializeMap = (coordinates: Coordinates) => {
-      if (!mapRef.current || !kakao?.maps) return;
+    const currentLevel = map.getLevel();
+    map.setLevel(currentLevel - 1);
+  };
 
-      const mapInstance = new kakao.maps.Map(mapRef.current, {
-        center: new kakao.maps.LatLng(coordinates.latitude, coordinates.longitude),
-        level: DEFAULT_ZOOM_LEVEL,
-      });
+  const zoomOut = () => {
+    if (!map) return;
 
-      setMap(mapInstance);
-      setZoom(mapInstance.getLevel());
-
-      kakao.maps.event.addListener(mapInstance, 'zoom_changed', () => {
-        setZoom(mapInstance.getLevel());
-      });
-    };
-
-    initializeMap(initialCoordinates);
-  }, []);
+    const currentLevel = map.getLevel();
+    map.setLevel(currentLevel + 1);
+  };
 
   return {
     map,
@@ -172,12 +124,11 @@ const useTreeMap = () => {
     centerPosition,
     bounds,
     zoom,
-    addMarker,
-    addCustomOverlay,
-    clearMarkers,
     updateCenterPosition,
     updateBounds,
     updatePosition,
+    zoomIn,
+    zoomOut,
   };
 };
