+++ frontend/src/pages/TreeMap/TreeMap.tsx
@@ -4,7 +4,7 @@ import DelayedButton from '@/components/_common/DelayedButton/DelayedButton';
 import FloatingButton from '@/components/_common/FloatingButton/FloatingButton';
 import Modal from '@/components/_common/Modal/Modal';
 import useModal from '@/hooks/_common/useModal';
-import useMapMarkers from '@/hooks/TreeMap/useMapMarker';
+import useMapMarkers from '@/hooks/TreeMap/useMapMarkers';
 import { useMapModal } from '@/hooks/TreeMap/useMapModal';
 import useTreeMap from '@/hooks/TreeMap/useTreeMap';
 import useTreeClustersQuery from '@/queries/Tree/useTreeClustersQuery';
@@ -19,7 +19,7 @@ const TreeMap = () => {
   const location = useLocation();
 
   const { isModalOpen, openModal, closeModal } = useModal();
-  const { map, mapRef, centerPosition, updatePosition, bounds, zoom } = useTreeMap();
+  const { map, mapRef, centerPosition, updatePosition, bounds, zoom, zoomIn } = useTreeMap();
 
   const isClusterView = zoom > CLUSTER_VIEW_THRESHOLD;
 
@@ -48,6 +48,7 @@ const TreeMap = () => {
     trees,
     treeClusters,
     onMarkerClick: handleMarkerClick,
+    onClusterClick: zoomIn,
   });
 
   const { modalType, modalContent, handleCloseModal } = useMapModal({
